+++ CHANGELOG.md
@@ -1,4 +1,15 @@
 # merakitools - CHANGELOG
+### v0.1.6
+ - [new] Create organization, claim orders, and add admins
+ - [new] Enable SAML and update SAML parameters
+ - [new] Accept organization name *or* ID
+ - [new] New spinner and table format
+
+ - [fix] *orgs api* command can list, enable, or disable and is now idempotent
+ - [fix] Python Black library for better formatting
+ - [fix] Find single organization without y/n questions
+ - [fix] various refactoring
+
 ### v0.1.5
  - Added multiple commands to **ms** and **mr** modules